requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-crontab==0.7.1
djangorestframework==3.14.0
Pillow==9.5.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2023.3
sqlparse==0.4.4
